Change all pointer loads to Acquire.
To be safe, all pointer loads are changed to Acquire, to be completely certain
that they cannot be dereferenced to a cached value from before the node was
allocated and constructed. Now all operations on AtomicPtr have Acquire and
Release semantics.

The current_height operations are left as Relaxed because they are just a
performance optimization: getting a stale height just makes the look up
slightly slower.

See discussion on Reddit:
https://old.reddit.com/r/rust/comments/c4jgg2/kudzu_concurrent_set_and_map_data_structures/ery3qvx/